feat(dedicated): deplace ipv6 feedback link to vrack page (#12314)
ref: MANAGER-14648
