+++ packages/manager/modules/ng-layout-helpers/src/list/list-layout.component.js
@@ -18,7 +18,6 @@ export default {
     formatters: '<?',
     topbarOptions: '<?',
     customizeColumnsMap: '<?',
-    surveyUrl: '<?',
   },
   controller,
   template,
