+++ packages/manager/modules/vrack/src/vrack.constant.js
@@ -4,8 +4,11 @@ export const SURVEY_LANGUAGES = {
 };
 
 export const BASE_URL_SURVEY = 'https://survey.ovh.com/index.php/863655?lang=';
+export const US_SURVEY_LINK =
+  'https://ovhcloud-us.limesurvey.net/945112?lang=en';
 
 export default {
   SURVEY_LANGUAGES,
   BASE_URL_SURVEY,
+  US_SURVEY_LINK,
 };
