+++ packages/manager/modules/vrack/src/vrack.routing.js
@@ -1,6 +1,5 @@
 import { ListLayoutHelper } from '@ovh-ux/manager-ng-layout-helpers';
 import { getVrackOrderUrl } from './vrack-order';
-import { SURVEY_LANGUAGES, BASE_URL_SURVEY } from './vrack.constant';
 
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('vrack.index', {
@@ -25,17 +24,6 @@ export default /* @ngInject */ ($stateProvider) => {
       staticResources: () => true,
       defaultFilterColumn: () => 'serviceName',
       header: /* @ngInject */ ($translate) => $translate.instant('vrack_title'),
-      surveyUrl: /* @ngInject */ (coreConfig) => {
-        const userLanguage = coreConfig.getUserLanguage();
-        const languageToUse = SURVEY_LANGUAGES.ALLOWED.includes(userLanguage)
-          ? userLanguage
-          : SURVEY_LANGUAGES.DEFAULT;
-        // Get user
-        const user = coreConfig.getUser();
-
-        // Build url for survey link
-        return `${BASE_URL_SURVEY}${languageToUse}&nic=${user.nichandle}`;
-      },
       customizableColumns: () => true,
       getServiceNameLink: /* @ngInject */ ($state) => ({
         serviceName: vrackId,
